handle some Ember weirdness
Because we can't be sure when we're on the config page (because of how
Ember handles routing) we need to run the background.js on all pages on
the inthe.am domain, checking every 500ms for the API key and only
showing the popup when we find it. In the future I'd like to hook the
Ember app itself and get a callback on render completed, but I don't
understand Ember internals well enough to do this yet.